вливание single_service в main

+++ src/main/java/org/javaacademy/carad/dto/AdvertisementDto.java
@@ -1,17 +1,15 @@
 package org.javaacademy.carad.dto;
 
-import lombok.NonNull;
-import lombok.Value;
-import org.javaacademy.carad.entity.Car;
-
+import java.math.BigDecimal;
 import java.time.LocalDate;
+import lombok.Value;
 
 @Value
 public class AdvertisementDto {
     int id;
     String brandName;
     String color;
-    double price;
+    BigDecimal price;
     String model;
     LocalDate localDate;
 }
